Fix blank screen issue by updating theme to use Material Light theme
The app was showing a blank white screen because the theme in themes.xml was
referencing Theme.AppCompat which wasn't available as a dependency. Updated to
use android:Theme.Material.Light.NoActionBar which is part of the Android SDK.

This fixes the UI rendering issue and the OnboardingScreen now displays correctly.